requirements.txt
aiogram==3.0.0
python-dotenv==1.0.0
yandex-music==2.1.1
requests==2.31.0
aiohttp==3.8.5
pydub==0.25.1
ffmpeg-python==0.2.0